Code restructuring: - Split some files to make our code easier to understand - Restructured our framework - Added a framework.dart library, to easily import all needed components for our framework - deleted the test file - simplified app startup
