+++ lib/framework/components/top_bar/top_bar.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:quiz/core/framework/top_bar/top_bar_cubit.dart';
+import 'package:quiz/framework/components/top_bar/top_bar_cubit.dart';
 import 'package:vrouter/vrouter.dart';
 
+/// The TopBar is our highly customized AppBar, which can be changed by using the [TopBarApi] via the [BuildContext.get] method.
 class TopBar extends StatelessWidget implements PreferredSizeWidget {
   @override
   Widget build(BuildContext context) {
