+++ lib/framework/components/view/abstract_view.dart
