+++ lib/framework/exceptions/stupid_developer_exception.dart
