+++ lib/framework/framework.dart
@@ -1,0 +1,20 @@
+library framework;
+
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:vrouter/vrouter.dart';
+
+import 'app/start.dart';
+import 'package:flutter/material.dart';
+
+import 'components/top_bar/top_bar_cubit.dart';
+import 'exceptions/stupid_developer_exception.dart';
+
+export 'package:flutter/material.dart';
+
+part 'app/application.dart';
+
+part 'api/api.dart';
+
+part 'components/top_bar/top_bar_api.dart';
+
+part 'router/router.dart';
